+++ packages/expo-router/src/testing-library/index.tsx
@@ -15,6 +15,10 @@ import { router } from '../imperative-api';
 // re-export everything
 export * from '@testing-library/react-native';
 
+afterAll(() => {
+  store.cleanup();
+});
+
 type RenderRouterOptions = Parameters<typeof render>[1] & {
   initialUrl?: any;
 };